changed driver lib/pq -> pgx/v5/pgxpool(it is better for async code) refactoring

+++ migrations/20221015144307_orders.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS orders(
     pk BIGSERIAL PRIMARY KEY,
     id VARCHAR(50), -- number in JSON obj
-    user_id INT REFERENCES users(id) ON DELETE CASCADE,
+    user_id BIGINT REFERENCES users(id) ON DELETE CASCADE,
     status VARCHAR(50),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     accrual FLOAT DEFAULT 0.0
